Moving serializers to seperate package, adding util files
