+++ MHacks/v1/serializers/__init__.py
@@ -1,0 +1,1 @@
+from MHacks.v1.serializers.serializers import *
